+++ app/src/main/kotlin/org/fossify/home/interfaces/HomeScreenGridItemsDao.kt
@@ -41,11 +41,18 @@ interface HomeScreenGridItemsDao {
         deleteItemsWithParentId(id)
     }
 
-    @Query("DELETE FROM home_screen_grid_items WHERE package_name = :packageName")
-    fun deleteItemByPackageName(packageName: String)
+    @Query(
+        "DELETE FROM home_screen_grid_items WHERE package_name = :packageName " +
+            "AND activity_name = :activityName AND user_serial = :userSerial"
+    )
+    fun deleteItemByIdentifier(packageName: String, activityName: String, userSerial: Long)
 
-    @Query("DELETE FROM home_screen_grid_items WHERE parent_id IN (SELECT id FROM home_screen_grid_items WHERE package_name = :packageName)")
-    fun deleteItemsByParentPackageName(packageName: String)
+    @Query(
+        "DELETE FROM home_screen_grid_items WHERE parent_id IN (" +
+            "SELECT id FROM home_screen_grid_items WHERE package_name = :packageName " +
+            "AND activity_name = :activityName AND user_serial = :userSerial)"
+    )
+    fun deleteItemsByParentIdentifier(packageName: String, activityName: String, userSerial: Long)
 
     @Query("UPDATE home_screen_grid_items SET `left` = `left` + :shiftBy WHERE parent_id == :folderId AND `left` > :shiftFrom AND id != :excludingId")
     fun shiftFolderItems(folderId: Long, shiftFrom: Int, shiftBy: Int, excludingId: Long? = null)
@@ -54,8 +61,8 @@ interface HomeScreenGridItemsDao {
     fun shiftPage(shiftFrom: Int, shiftBy: Int)
 
     @Transaction
-    fun deleteByPackageName(packageName: String) {
-        deleteItemByPackageName(packageName)
-        deleteItemsByParentPackageName(packageName)
+    fun deleteByIdentifier(packageName: String, activityName: String, userSerial: Long) {
+        deleteItemByIdentifier(packageName, activityName, userSerial)
+        deleteItemsByParentIdentifier(packageName, activityName, userSerial)
     }
 }